Cadastro da Pauta, Sessão e Voto; Criaddo controller para cada evento; Criado services da sessão para validar sessões abertas para votação e cadastro; Criado services para verificar CPF do associado, impedir que o associado realize mais de um voto na mesma sessão; Utilização do ModelMapper para converter entidades em Dtos, com criação do conversores para data;

+++ src/main/java/br/com/votacao/repository/PautaRepository.java
@@ -3,5 +3,5 @@ package br.com.votacao.repository;
 import br.com.votacao.domain.Pauta;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PautaRepository extends CrudRepository<Pauta, Integer> {
+public interface PautaRepository extends CrudRepository<Pauta, Long> {
 }
